requirements.txt
pandas
numpy
scikit-learn
tensorflow
keras
ccxt
pycoingecko
python-dotenv
taccxt 
